function_kompas_crawler/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
scrapy
requests
bs4
scrapydo